First attempt at implementing fit with datasets.
The previous data formats still work as usual.

To fit with a train dataset, just pass it as X, y must be a dummy or
None.

To fit with train and valid, pass a tuple of (dataset_train,
dataset_valid). For scoring still to work, we need to pass (y_train,
y_valid) as well, but this is only until scoring re-uses data from
batches.

There are not tests yet, since this is only a proposal.