Implement fitting and predicting with datasets.
It is now possible to fit and predict(_proba) with datasets. This is
especially useful if you use predefined datasets (say from
torchvision) or if you would like to load data lazily. Unfortunately,
when working with datasets, some sklearn-related features may not work
anymore. The Advanced_Usage.ipynb shows what can be done about it.