Cached scoring callbacks (#171)
Scoring callbacks now cache targets and predictions generated during a training epoch.

The cache can be used to efficiently determine scores on epoch end. Previously, each 
scoring callback (except train_loss and valid_loss) needed to make a separate prediction, 
which slows down training. It is possible to disable caching to revert to the old behavior 
(needs to be done on all scoring callbacks if memory is a concern).

The cache is cleaned automatically when training ends but this can be triggered 
manually be re-initializing the callbacks (or the whole net).

As an additional advantage, caching combines well with custom datasets, since using 
those would usually break scoring.