Add support for 'event_' columns in history
PrintLog will output such columns on the right, just before 'dur'. When
the value stored in history is True, a plus sign will be printed to the
table. When the value stored in history is False or None, nothing will
be printed, i.e. the cell will be empty.

The 'event_' prefix is stripped off from the key so as to not make the
table header unnecessary long.