Add support for 'event_' columns in history (#274)
* Add support for 'event_' columns in history

PrintLog will output such columns on the right, just before 'dur'. When
the value stored in history is True, a plus sign will be printed to the
table. When the value stored in history is False or None, nothing will
be printed, i.e. the cell will be empty.

The 'event_' prefix is stripped off from the key so as to not make the
table header unnecessary long.

* Add information about 'event_' columns to the PrintLog docstring

* Update Checkpoint callback

* Writes status to `event_cp` column;
* Default sink is set to noop.

* Fix function names in history documentation

* Add stralign parameter defaulting to 'right'