Fix a number issues with ProgressBar (#315)
Numerous fixes in ProgressBar

* Avoid excessive progress bar refreshing

Currently set_postfix() unconditionally prints the current progress bar
to the output stream at the end of every batch. This is not necessary
since it is followed by an update() call.

* Hide progress bar after the epoch is over

Currently the (completed) progress bar stays in the console output after
the epoch is over, screwing the logger table. This commit disables the
leave flag to make sure that the progress bar is wiped out.

* Fix "auto" mode in ProgressBar

* Set default batches_per_epoch to 'auto' and update docstrings