Refactor epoch scoring easier subclassing (#341)
* Minor refactoring of EpochScoring for easier subclassing.

With the old implementation, a lot of code needs to be duplicated if
on_epoch_end is overriden, making it difficult for users to implement
their own EpochScoring. This is greatly facilitated now.

* Move functionality to pick train/valid to convenience method.

This avoids possible confusion when caching is used, e.g. when the
on_train=False, but a user passes dataset_train, they still get cache
from dataset_valid, which can cause bugs.