Feature/better user facing messages (#374)
* Bugfix: Reinitialize optimizer when set_params is called with lr.

Previously, when the net was already initialized, calling
set_params(lr=new_lr) didn't trigger a re-initialization because the
method only looked for optimizer__* parameters. However, lr is
indirectly also an optimizer parameter, so setting it must
reinitialize the optimizer.

* Replace placeholder error message with useful one.

This is a safeguard that should usually never be reached but still
there should be a useful error message.

* Remove unnecessary import.

* Improve user-facing messages when re-initialization occurs.

Specifically, alert the user to the parameters that were responsible
for re-initialization. As a by-product, this change also prevents the
optimizer from being re-initialized twice when both it and the module
changed.

* Refactor code that creates user-facing message.

Also, improve docstrings.

* Add entry to CHANGES.md.