Add support for scipy sparse CSR matrices as input data. (#406)
Among other use cases, this allows to use transformers that return
sparse matrices in the model pipeline, e.g. sklearn's
CountVectorizer. Previously, data needed to be cast to dense arrays
before passing it to skorch.

Note that during batching, data is cast to dense because pytorch's
DataLoader cannot deal with sparse matrices. But even if it could, it
would be of little use because torch support for sparse arrays is very
incomplete at this time (e.g. nn.Linear does not work in
backwards). Still this allows to avoid having to cast a potentially
huge input sparse matrix to dense so the feature should be useful.

This PR also updates to_tensor to allow casting a scipy sparse matrix
to a torch sparse COO tensor when accept_sparse=True. Even though
skorch does not use this feature, it seemed to be useful for this
function in general, since with pytorch 1.0, sparse support became
better overall.