Remove deprecations for 0.6.0 release (#476)
* Remove deprecated f argument from save_params

Use f_params instead.

* Remove deprecated save_history and load_history

Now use save_params with f_history keyword.

* Remove deprecated filter_requires_grad and filtered_optimizers

Now use skorch.callbacks.Freezer if you want to freeze parameters.

* Remove deprecated "wrong" slice of history

It used to be possible to do history[j, 'batches', somekey] but that
was inconsistent and should really be history[j, 'batches', :,
somekey] because the 2nd to last key relates to the batch indices, not
the batch keys. This inconsistent behavior is no longer supported.

* Remove support for unpickling old skorch nets

With PyTorch 1.1.0, we can no longer unpickle old models that have an
optimizer (i.e. basically all models). Therefore, we may as well drop
the support for unpickling old files now, when moving to PyTorch
1.1.0.

* Remove target argument from Checkpoint

Use f_params instead.

* Remove deprecated device argument from Dataset

This argument didn't do anything. Now passing it raises an error.

* Change TrainEndCheckpoint fn_prefix default to 'train_end_'

It used to be 'final_'.

* Bump up PyTorch version to 1.1.0 in travis.yml

We need it for the CyclicLR implemenation in PyTorch.