Preparation for more correct get params change (#527)
* Introduce changes and tests in preparation for new get_params

The new behavior of get_params will be to not returned any "learned"
attributes such as "module_".

This PR implements the new behavior but doesn't switch to it yet to
give users time to adjust their code. This is a breaking change but it
is necessary since it is the "correct" behavior; the old one could
introduce subtle bugs in rare situations (e.g. `GridSearchCV` with a
net that has `warm_start=True`).

The PR also includes tests that are currently failing but that are
passing under the new behavior. When switching to the new behavior,
all tests, including these new ones, should pass (they currently
xfail).

* Make net.history a property that refers to net.history_

Add setter and getter methods for net.history. That way, history now
ends on '_' like all other parameters that are not provided directly
by the user.

Co-Authored-By: ottonemo <marian.tietz@ottogroup.com>