Don't iterate over data when using cached scoring
Before, net.infer was cached when using a scoring callback wiht
use_caching=True. This way, the time to make an inference step was
saved. However, there was still an iteration step over the data for
each scoring callback. If iteration is slow, this could incur a
significant overhead.

Now net.forward_iter is cached instead. This way, the iteration over
the data is skipped and the iteration overhead should be gone.