Fix a bug with custom attribute name clashes
When a user adds a new "settable" attribute (i.e. that works with
set_params) whose names starts the same as an existing attribute (say
"optimizer_2"), and adds a corresponding argument (say
"optimizer_2__lr"), skorch will erroneously complain about this
argument because it thinks it belongs to the other
attribute ("optimizer" in this example). The added unit test should
illustrate this behavior.