Introduce PassthroughScoring
A scorer that just passes through a score that is already calculated
and logged during the batches (as, e.g., train_loss). Replaces use of
BatchScoring for most use cases.

This makes it a lot easier to add a custom loss during a batch. Just
log it in the history at each batch and add PassthroughScoring with
the name of the key of that loss. No need to add a fake scoring
function that just returns the value.