Make get_params_for and get_params_for_optimizer public
Private methods are kept for compatibility. Functionally, nothing was
changed.

The main reason to do this is to give users confidence in, and better
documentation for, implementing their own `initialize_` methods. Those
methods require the ues of `get_params_for`, which until now was
discouraged since it used to be private.

One example where this is useful is if the user wants to implement a
net that has more than one optimizer. Now they can override
`initialize_optimizer` and use `get_params_for_optimizer` to
initialize all their optimizers.