Refactor tests that used patched score method
The score method used to be patched to give fixed values for a net
fixture used variously in the tests. This was so that certain scores
could be asserted. However, this had the side effect that some tests
that actually want to calculate the real score had mysterious return
values from score().

Instead of fixing those tests, it's better not to patch the score
method in the fixture and to instead patch it in those tests where
it's necessary. This helps making those tests more explicit (which is
better than implicit).