Facilitate adding custom modules, criteria, or optimizers
Now, when creating a custom net that has additional modules, criteria,
or optimizers, skorch will take care of registering them inside
`prefixes_` and `cuda_dependent_attributes_`. If this non-obvious
change is not done, it can result in errors down the line.

Additionally, during `set_params`, there is now re-initialization of
those components if any of the special keys (i.e. those that start
with one of the `prefixes_`) has "module", "criterion", or "optimizer"
as a substring.

This last part is not a perfect solution, but I couldn't find a better
way to determine whether a component needs to be reinitialized. Say we
add a new module called "mymodule" to the custom net. How can we know
if calling `net.set_params(mymodule__foo=123)` should result in module
re-initialization? We can't really, except by trying to guess from the
name. Other ideas that I thought of but rejected:

1. Check if `self.module` is a `nn.Module` subclass. But what if it's
not a module subclass but a function that returns an instantiated
module?

2. Instantiate `self.mymodule()` and see if we get a module back. But
what if the module requires arguments in order to be called
successfully? Okay, we could find those arguments and then
initialize. But do we really want to make a "test initialization"? If
feel like this could have unintended side-effects.

If anybody can propose a better idea, that would be welcome. My
personal belief is that the substring matching is not that bad in the
sense that it's very unlikely to trigger a false positive. However, it
might not be easy to discover for a user (I added some documenation
and plan on adding the DCGAN example that shows this in practice).