Switch from old get_params to new get_params behavior
get_params no longer returns learnable parameters like module_. This
is in line with what sklearn expects but can potentially break old
code that relied on it implicitly. Since get_params is often uesd
under the hood in sklearn (e.g. by GridSearchCV), this could affect a
lot of code.

Still this new behavior is better so we decided to go along with
it. We had an announcement for 0.7 that warned about this. Now it's
coming in 0.9 (probably 1 year later or so, thus enough time to
adapt).