Change: new get params behavior and bugfix (#617)
* Switch from old get_params to new get_params behavior

get_params no longer returns learnable parameters like module_. This
is in line with what sklearn expects but can potentially break old
code that relied on it implicitly. Since get_params is often uesd
under the hood in sklearn (e.g. by GridSearchCV), this could affect a
lot of code.

Still this new behavior is better so we decided to go along with
it. We had an announcement for 0.7 that warned about this. Now it's
coming in 0.9 (probably 1 year later or so, thus enough time to
adapt).

* Fix the random seed for tests inside conftest

Use an autouse fixture so that all our tests are now using the same
random seed. This is better than fixing the seed at the beginning of
the module (which is the previous implementation) because it results
in more consistent behavior when only running a subset of all tests.

* Fix a bug in tests with fixtures referencing same net

We had fixtures net and net_fit. However, they both referenced the
same object. Therefore, after doing net.set_params, we actually
mutated net_fit as well. This is no longer happening.

Note that I first wanted to do net_fit = clone(net).fit(X,y).
However, clone(net) was no good because there is a callback defined as
Mock(spec=Callback) that, when cloned, turns into Callback. This is
not desired, so I turned to creating a completely new instance for
net_fit.