Fix a TypeError that occurred when using predefined_split
This bug was introduced by PR 646 after a change in the function
signature of utils._make_split. Here the bug is fixed by returning to
the old signature but giving default values for valid_ds and y, so
that the desired behavior introduced by 646 is still possible.

A test is introduced to catch this bug.

As a cleanup, the unrelated test_predefined_split now uses the
already defined dataset_cls fixture instead of importing Dataset again.