Make cli helper work with normal sklearn estimators
Change cli helper so that normal (i.e. non-skorch) sklearn estimators
also work. This mainly required to relax the assumption that the
estimator must have a 'module' attribute.

On top, added a comment about not using positional arguments for
functions parsed by fire. The problem is that internally,
this (silently) breaks fire, so that 'python foo.py --help' does not
work and instead returns the result of 'python foo.py -- --help'.