GradientNormClipping checks if optimizer is unscaled
Since unscaling is not an idempotent operation, PyTorch raises a
RuntimeError if an unscaled optimizer is unscaled again. Since we
could have multiple callbacks or other components that might want to
unscale, we have to protect against this possibility. Therefore, we
check if the optimizer has already been unscaled before unscaling.

Unfortunately, we have to use a privat attribute in PyTorch for this,
so this is prone to break in the future.