Add a logging callback for Sacred
Similar to the existing Neptune logger, just logging to Sacred instead
of Neptune. The main difference is the name of the logging function and
that an experiment does not have to be closed explicitly. I kept the
documentation and functionality similar, to make it easier to use one as
a drop-in replacement for the other.