More efficient history access
Changes implementation of History.__getitem__ so that the type of
items is only looked up once, not each time for each element. This
should be fine, as the types should not change between
elements (e.g. the rows in 'batches').

This can save a lot of time. In an example using 1000 samples, 500
epochs and a very small net, history access time goes down from 10 sec
to 2.5 sec.