Refactor train loop for easier customization (#699)
* Small refactoring of some training/evaluation methods

The main change is to "push down" the unpacking of the batch
data (using "Xi, yi = unpack_data(batch)") from "run_single_epoch" to
functions lower down the chain. The reason for this is that if we put
this unpacking rather "high up" the chain, we force the output of
dataset to be a tuple of 2 elements. If a user wants to change this,
they either have to modify their dataset to alwas return 2
elments (which can be annoying) or they would need to override
"run_single_epoch", but we don't want to encourage this, as this
method performs a lot of book keeping.

With the new changes, small methods further down the chain are now
responsible for dealing with the data, e.g. "train_step_single" and
"validation_step". This affords more flexibility overall, and a user
can easily override these methods without having to deal with book
keeping.

Obviously, this change results in some API changes, so some existing
code might break as a consequence (most notably code that relies on
"on_batch_begin" or "on_batch_end").

Another "victim" of this change is the checking for whether y is a
placeholder and then replacing it with None. Tbh. I found this to be
an ugly solution anyway. Now it can happen that when a user uses the
skorch Dataset, they will receive a 0s tensor for y when they set y to
None. I think this is an acceptable loss overall.

To illustrate the gains of the change, I added a test,
"test_customize_net_with_custom_dataset_that_returns_3_values", which
shows what is now possible when it comes to customizing skorch. With
the old state, achieving the same would have been more cumbersome.

Finally, I added a section to the docs called "Customization". The
intent is to help the user decide how best to customize "NeuralNet" et
al. It explains what methods best to override for what purpose and
what methods to best leave untouched. In the future, we could expand
this section to cover more aspects.

* Clean ups to satisfy pylint

* Add more to .gitignore and remove duplication from .coveragerc

* Clarify doc for return value of train_step

Also, add documentation of return value of train_step_single, which
was missing (same docstring).

* Reviewer comment: avoid unnecessary lambda in test

Co-authored-by: Thomas J. Fan <thomasjpfan@gmail.com>

* Add migration guide for affected methods to FAQ

As suggested by reviewer.

* Reviewer comment: tuple unpacking

Co-authored-by: githubnemo <githubnemo@users.noreply.github.com>

* Reviewer comment: tuple unpacking

Co-authored-by: githubnemo <githubnemo@users.noreply.github.com>

* Add a helpful warning if user-defined on_batch

Check if the user uses any custom callback that overrides on_batch_begin
or on_batch_end. If they do, warn about the signature change and explain
how to recover. Also, indicate how to suppress the warning.

* Add tests for warning when on_batch overridden

Forgot to check this in when the warning was added.

Co-authored-by: BenjaminBossan <b.bossan@gmail.com>
Co-authored-by: Thomas J. Fan <thomasjpfan@gmail.com>
Co-authored-by: githubnemo <githubnemo@users.noreply.github.com>