Further updates based on new skorch refactoring
Merge in latest branch of the refactoring that elevates custom modules
to first class citizens.

Moreover, provide an ExactGPRegressor class directly, instead of
requiring the user to implement this. That way, we avoid the minor
annoyance of having to write your own initialize_module.

Consequently, the notebook could be simplified and the futureattr
workaround could be removed.

Furthermore, add a lost of tests, most notably tests for
GPRegressor (not exact) and GPBinaryClassifier.

Added some tests for grid search as well, which showed a few issues that
were addressed.