Fix a bug when likelihood/module are initialized
When the GP is initialized with likelihood/module that are already
intialized, it's not required to create new instances if parameters
haven't changed. The logic for this was not quite correct but has been
corrected now.

Testing this for modules was not possible because all GP modules require
params, thus we cannot construct an example without module params.