Implement loading of checkpoints after training (#754)
In conjunction with early stopping (and particularly grid searches)
this feature can come in handy as you would certainly want to use the
best model state for scoring your model after training without worrying
about loading it from the checkpoint manually.

It would be possible to implement this in other places, such as the
early stopping callback. To me this seems to be a conflict of
separation, which is why I chose this place.