Gaussian processes via gpytorch (#782)
This is a feature to add support for Gaussian Processes (GPs) via integration
with gpytorch.

Similar to "vanilla" PyTorch, the idea here is that skorch allows the user to
focus on what's important (implementing the mean function and kernel function)
and not to bother with stuff like the training loop, callbacks, etc. This is
probably best illustrated in the accompanying notebook.

GPs are primarily for regression, hence those are the main focus here.
Traditionally, there are "exact" solutions and approximations. skorch will
provide an ExactGPRegressor and a GPRegressor for those two use cases.

On top of that, a GPBinaryClassifier is offered, though I suspect it to be
rarely used. I couldn't get the GPClassifier for multiclass to work, the code is
therefore commented out.

The API is mostly the same as for the normal skorch estimators. There are some
additions to make working with GPs easier:

- predict method takes a return_std argument to return the standard deviation as
  well (as in sklearn's GaussianProcessRegressor; return_cov is not supported)
- sample method to sample for the model
- confidence_region method to get the confidence region

Co-authored-by: Thomas J. Fan <thomasjpfan@gmail.com>