Automatically use accelerator.print by default
This is achieved by setting the default callbacks__print_log__print to
'auto'. When this is detected later, replace it with
self.accelerator.print if available.

This way, we can use the sane default even if we cannot directly set it
as default (because the accelerator instance does not exist yet) but
still give the user the option to set a different argument.