Fix a bug caused by using closure with GradScaler
When using AMP, accelerate applies grad scaling under the hood using
GradScaler. That does not support passing the train step as a closure to
optimizer.step. Therefore, we need to step explicitly.

We could use a more sophisticated approach of trying to figure out if
grad scaler is actually being used and only stepping explicitly if
needed. However, the need for the closure is quite rare and we want to
treat accelerate as a black box instead of relying on implementation
details (which we would have to in order to figure out when grad scaling
is applied).