Facilitate integration with huggingface
Solves 838

Description

The Huggingface transformers library (or, more precisely, the tokenizers
library) currently is possibly the most popular library for state of the
art NLP models. Therefore, it would be good if skorch works well with
transformers.

One of the key steps is tokenization. For this, tokenizers implements
their own classes that return a special data type, which is currently
not understood by skorch. This PR makes it so that skorch understands
and thus correctly works with that datatype. Without this adjustment,
users would need to write some (light) wrapper code to make skorch work
with these tokenizers.

Implementation

The problem with this special data type, BatchEncoding, is that it's a
Mapping but not a dict. skorch already deals with dicts but since
Mappings are not instances of dicts, skorch does not recognize the
BatchEncoding correctly. Therefore, some isinstance checks had to be
adjusted.

The second difficulty with BatchEncodings is that their values are lists
of lists (e.g. lists of lists of ints for the token ids). This clashes
with skorch which, when encountering a list, believes that the list
contains the actual values. Therefore, when we determine the length of a
BatchEncoding, we don't get the length of the outer list, which is the
desired value (the sample size), but we get the length of the inner
list (the sequence length). This in turn leads to an error with our
Dataset, which believes that lengths are inconsistent.

To work around this issue, we could change how the get_len
implementation deals with lists of values. However, this would introduce
a backwards incompatibility. Thus I took a different, somewhat
inefficient approach: I call the convert_to_tensors method on
BatchEncoding, which converts the values from lists of lists to (padded)
tensors. For those, get_len returns the correct results.

This conversion is obviously inefficient. Hopefully, however, this makes
no big difference overall because the Dataset is initialized only once
per fit call.