Reviewer suggestion: get length of BatchEncoding
The previous implementation performed a conversion to tensors, which may
be expensive and unnecessary in some situations. Instead, now directly
return the length of the 'input_ids' list if it is found.

Instead of checking for the method convert_to_tensors, we now check if
the data is a Mapping and if it has the 'input_ids' key.