Rework HuggingfaceTokenizer
First of all, similar to the NeuralNet class, it is now possible to pass
uninitialized classes and have them initialized before fit. Most of the
machinery is a copy (with adjustments) of the same code form NeuralNet.

With these changes, it is now possible to perform grid searches etc. on
the individual parameters (instead of having to pass many instances).

Furthermore, by default the trainer is now being inferred. Thus it no
longer needs to be passed by the user and stored as an attribute on the
estimator. This mostly circumvents the pickling issues as well.