Add function to trim net for prediction
After finishing training, a net has many attributes that are not needed
anymore when only inference is performed. A user might want to get rid
of these attributes to keep the size of the net small. E.g., depending
on the chosen optimizer, it can be quite big.

(In addition, there is the advantage that the less attributes exist, the
more likely a net can be unpickled that used a different version of
skorch, PyTorch, etc.).

This commit adds a convenience function, trim_for_prediction, that takes
care of removing all attributes that are no longer required.

Implementation

In addition to removing unneeded attributes, make sure to clear
callbacks (which, at least at the moment, are used exclusively during
training), the history, and the train_split. Training is set to False.

Also, set an attribute after trimming. When the net is
initialized/fitted, check if the net is trimmed and raise a useful error
message for the user. This check does not require the attribute to be
set, to prevent possible compatibility issues. For this, a new training
readiness check and exception are introduced.

If someone can think of a better name for 'trim_for_prediction', feel
free to suggest.

An open question: Do we want to provide this functionality as an
argument for TrainEndCheckpoint? If so, just as True/False or as a
callable?