Facilitate integration with huggingface (#840)
Description

The Huggingface transformers library (or, more precisely, the tokenizers
library) currently is possibly the most popular library for state of the
art NLP models. Therefore, it would be good if skorch works well with
transformers.

One of the key steps is tokenization. For this, tokenizers implements
their own classes that return a special data type, which is currently
not understood by skorch. This PR makes it so that skorch understands
and thus correctly works with that datatype. Without this adjustment,
users would need to write some (light) wrapper code to make skorch work
with these tokenizers.

Implementation

The problem with this special data type, BatchEncoding, is that it's a
Mapping but not a dict. skorch already deals with dicts but since
Mappings are not instances of dicts, skorch does not recognize the
BatchEncoding correctly. Therefore, some isinstance checks had to be
adjusted.

The second difficulty with BatchEncodings is that their values are lists
of lists (e.g. lists of lists of ints for the token ids). This clashes
with skorch which, when encountering a list, believes that the list
contains the actual values. Therefore, when we determine the length of a
BatchEncoding, we now check if the data is a Mapping and if it has the
'input_ids' key, in which case we use the length of that element as the
sample size.

Co-authored-by: ottonemo <marian.tietz@ottogroup.com>