Bugfix: sklearn 1.1 and SliceDataset (#858)
Change in sklearn error message

The error message when using set_params with an invalid key has been
altered to contain quote marks. Now testing message without or with.

Fix issue with SliceDataset

Using SliceDataset could result in an error because of this line:

self.classes_inferred_ = np.unique(to_numpy(y))

This would fail with a SliceDataset. However, the error was masked
because GridSearchCV swallows errors. Therefore, we now run these tests
with error='raise' to surface the bug.

Regarding the bugfix, it is a bit clumsy because we need to make
to_numpy work with SliceDataset, but we cannot check

'isinstance(X, SliceDataset)'

because we don't want to import the helper class. Therefore, we now
check this indirectly by looking at attributes.

Furthermore, SliceDataset now also works with torch tensors as X and y,
not only numpy arrays. Tests have been extended to cover this. Also, it
now supports np.asarray.