torch geometric support + example notebook
Adding support for torch geometric has so far been straight-forward
since its data types also implement the `.to()` interface and use
tensors in the background, they can basically be passed forward to
the module which will deal with the data type accordingly.

There are probably other places that might still fail, I can imagine
that different data loaders (there are quite a [few][1]) might be
problematic in terms of splitting X/y in batches, but this is a
first step forward and we can address these issues once they arise.

[1]: https://pytorch-geometric.readthedocs.io/en/latest/modules/loader.html