Address reviewer comments
- When vocabulary_ is not set, raise AttributeError now use
- get_feature_names_out instead of the deprecatred get_feature_names
- implement __sklearn_is_fitted__ method for check_is_fitted
- remove unnecessary getattr use

Incidental change:

skorch's check_is_fitted function now makes use of sklearn's
check_is_fitted function. It intercepts the sklearn error and raises a
skorch error instead. This is already in line with the comment on this
function which claimed that the sklearn function is being used under the
hood, even though that was not true.